+++ kedro_neptune/__init__.py
@@ -409,12 +409,8 @@ class NeptuneHooks:
 
     # pylint: disable=unused-argument
     @hook_impl
-    def after_catalog_created(
-            self,
-            catalog: DataCatalog,
-            save_version: str,
-    ) -> None:
-        self._run_id = hashlib.md5(save_version.encode()).hexdigest()
+    def after_catalog_created(self, catalog: DataCatalog) -> None:
+        self._run_id = hashlib.md5(str(time.time()).encode()).hexdigest()
         os.environ['NEPTUNE_CUSTOM_RUN_ID'] = self._run_id
 
         catalog.add(