+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [UNRELEASED] kedro-neptune 0.1.2
+
+### Fixes
+-  Fixed custom run id generation on catalog creation ([#42](https://github.com/neptune-ai/kedro-neptune/pull/42))
+
 ## kedro-neptune 0.1.1
 
 ### Features